requirements.txt
flask==3.0.2
python-dotenv==1.0.1
pytest==8.1.1
requests==2.31.0
python-jose==3.3.0
anthropic==0.18.1
openai==1.13.3
mistralai==0.0.12